requirements.txt
aiohttp==3.6.3
alembic==1.5.5
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.3.9
constantly==15.1.0
cryptography==3.3.1
cssselect2==0.4.1
dateparser==1.0.0
defusedxml==0.7.1
discord.py==1.5.1
distlib==0.3.1
filelock==3.0.12
firebase-admin==4.4.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
google-api-core==1.23.0
google-api-python-client==1.12.6
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10
imgkit==1.0.2
importlib-metadata==3.4.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.3
Mako==1.1.4
MarkupSafe==1.1.1
mimerender==0.6.0
msgpack==1.0.0
multidict==4.7.6
pdfkit==0.6.1
Pillow==7.1.2
proto-plus==1.11.0
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-barcode==0.11.0
python-binance==0.7.5
python-dateutil==2.8.1
python-editor==1.0.4
python-mimeparse==1.6.0
pytz==2020.4
pyzbar==0.1.8
regex==2020.11.13
reportlab==3.5.68
requests==2.25.0
rsa==4.6
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
svglib==1.1.0
tinycss2==1.1.0
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.2
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.5.1
zipp==3.4.0
zope.interface==5.2.0
